upgrade from sf 5.4 to 6.4
* upgrade from sf 5.4 to 6.4

* fix deprecated

* fix remember me always activated bug

---------

+++ src/Controller/UserController.php
@@ -4,20 +4,20 @@ namespace App\Controller;
 
 use App\Form\UserType;
 use Doctrine\ORM\EntityManagerInterface;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
-#[IsGranted('ROLE_USER')]
 #[Route('/my-account')]
 class UserController extends AbstractController
 {
     #[Route('/', name: 'profile', methods: ['GET'])]
     public function show(): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
         $user = $this->getUser();
 
         return $this->render('user/show.html.twig', [
@@ -28,6 +28,7 @@ class UserController extends AbstractController
     #[Route('/edit', name: 'edit_user', methods: ['GET', 'POST'])]
     public function edit(Request $request, EntityManagerInterface $entityManager, UserPasswordHasherInterface $userPasswordHasher): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
         $user = $this->getUser();
 
         $form = $this->createForm(UserType::class, $user);
@@ -50,22 +51,23 @@ class UserController extends AbstractController
             return $this->redirectToRoute('profile', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->renderForm('user/edit.html.twig', [
+        return $this->render('user/edit.html.twig', [
             'user' => $user,
             'form' => $form,
         ]);
     }
 
     #[Route('/delete', name: 'delete_user', methods: ['POST'])]
-    public function delete(Request $request, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, EntityManagerInterface $entityManager, TokenStorageInterface $tokenStorage): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
         $user = $this->getUser();
 
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_csrf_token'))) {
             $entityManager->remove($user);
             $entityManager->flush();
 
-            $this->container->get('security.token_storage')->setToken(null);
+            $tokenStorage->setToken(null);
         }
 
         return $this->redirectToRoute('home', [], Response::HTTP_SEE_OTHER);
